Add initial limine.yml configuration for aarch64 architecture
This commit introduces a new limine.yml file for the aarch64 architecture, specifying the target triple as 'arm64-apple-darwin' and setting the binary path to 'limine'. The relocations section is currently empty.

+++ src/kernel/panic.c
@@ -12,7 +12,7 @@ void panic(const char *message) {
     for (;;) { __asm__ volatile("cli; hlt"); }
 }
 
-void panic_assert(const char *file, uint32_t line, const char *desc) {
+void panic_assert(const char *file, uint64_t line, const char *desc) {
     __asm__ volatile("cli");
     serial_puts("\n=== ASSERTION FAILED ===\nFile: ");
     serial_puts(file);
